Fixed MSVC warnings and added ico7.obj

+++ tp2/projets/meshIOUtils.cpp
@@ -6,7 +6,7 @@ std::vector<Triangle> MeshIOUtils::create_triangles(MeshIOData meshData, Transfo
 	std::vector<Triangle> _triangles;
 	_triangles.reserve(meshData.indices.size() / 3);
 
-	for (unsigned int i = 0; i < meshData.indices.size(); i += 3)
+	for (size_t i = 0; i < meshData.indices.size(); i += 3)
 	{
 		int index0 = meshData.indices[i + 0];
 		int index1 = meshData.indices[i + 1];
